requirements.txt
pandas==2.2.2
requests==2.32.3
openpyxl==3.1.5
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pyyaml==6.0.2
dbt-postgres==1.8.2
